requirements.txt
# Core runtime dependencies for Local-Video-RAG
fastapi>=0.95.2
uvicorn[standard]>=0.22.0
pandas>=1.5.3
numpy>=1.24.3
scikit-learn>=1.2.2
requests>=2.28.2
python-dotenv>=1.0.0

# Transcription pipeline
openai-whisper>=20230314
# PyTorch: install the appropriate wheel for your platform (CPU or CUDA) — see README.md
torch

# Phase 2: Vector Database
chromadb>=0.4.0
